change: Move terminal colors into chat.py
This is to fix import errors when running the script from main.
