--- src/modules/terminal_colors.py
@@ -1,5 +1,0 @@
-class t_color:
-    RED = '\033[31m'
-    GREEN = '\033[32m'
-    BLUE = '\033[34m'
-    RESET = '\033[0m'
